# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Jarbas Remote Executor Bridge",
    "description": "API para controlar remotamente um computador Windows. Permite executar comandos no PowerShell, Codex, Claude CLI e outras ferramentas, além de operações completas no sistema de arquivos (ler, escrever, listar, deletar, editar). O bridge roda no PC local e é acessado via Cloudflare Tunnel.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://presentations-rest-studios-tab.trycloudflare.com",
      "description": "Bridge server via Cloudflare Tunnel (URL atualizada automaticamente)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token de autenticação configurado como BRIDGE_TOKEN no servidor"
      }
    },
    "schemas": {
...
      "FileEntry": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "package.json" },
          "path": { "type": "string", "example": "D:\\jarbas_vida\\package.json" },
          "type": { "type": "string", "enum": ["file", "directory"] },
          "size": { "type": "integer", "example": 512 },
          "modified": { "type": "string", "format": "date-time" },
          "children": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/FileEntry" },
            "description": "Subentradas (só aparece com recursive: true)"
          }
        }
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
  "paths": {
...
    "/run": {
      "post": {
        "summary": "Executar comando no PC remoto",
        "description": "Executa um comando usando uma ferramenta autorizada no computador Windows. Use 'powershell' para comandos do sistema (Get-Process, Get-Service, Install-Module, etc.), 'codex' para Claude Code, 'claude' para Claude CLI, 'python' para scripts Python, 'node' para Node.js. O PowerShell roda com ExecutionPolicy Bypass para acesso admin completo.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tool", "command"],
                "properties": {
                  "tool": {
                    "type": "string",
                    "description": "Ferramenta a usar (deve estar na allowlist)",
                    "enum": ["powershell", "codex", "claude", "python", "node"],
                    "example": "powershell"
                  },
                  "command": {
                    "type": "string",
                    "description": "Comando a executar",
                    "example": "Get-Process | Select-Object -First 5 Name, CPU, WorkingSet"
                  },
                  "args": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Argumentos adicionais (para tools que não são powershell)",
                    "example": []
                  },
                  "timeout": {
                    "type": "integer",
                    "description": "Timeout em ms (padrão: 300000 = 5min)",
                    "example": 60000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Comando executado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "stdout": { "type": "string", "description": "Saída padrão do comando" },
                    "stderr": { "type": "string", "description": "Saída de erro do comando" },
                    "exitCode": { "type": "integer", "example": 0 },
                    "executionTime": { "type": "integer", "description": "Tempo de execução em ms", "example": 1234 }
                  }
                }
              }
            }
          },
          "400": { "description": "Campos obrigatórios faltando (tool, command)" },
          "401": { "description": "Token de autenticação inválido ou ausente" },
          "403": { "description": "Ferramenta não está na allowlist" },
          "500": { "description": "Comando falhou na execução" }
        }
      }
    },
    "/jobs/run": {
      "post": {
        "summary": "Executar comando assíncrono (retorna jobId imediatamente)",
        "description": "Cria um job assíncrono para comandos demorados como codex, claude, etc. Retorna um jobId imediatamente. Use GET /jobs/{jobId} para consultar o resultado depois. Ideal para comandos que podem levar mais de 30 segundos.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["tool", "command"],
                "properties": {
                  "tool": {
                    "type": "string",
                    "description": "Ferramenta a usar",
                    "example": "codex"
                  },
                  "command": {
                    "type": "string",
                    "description": "Comando a executar",
                    "example": "liste os arquivos do projeto"
                  },
                  "args": {
                    "type": "array",
                    "items": { "type": "string" },
                    "description": "Argumentos adicionais"
                  },
                  "timeout": {
                    "type": "integer",
                    "description": "Timeout em ms (padrão: 600000 = 10min)",
                    "example": 300000
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "jobId": { "type": "string", "example": "job_1738850000000_1" },
                    "status": { "type": "string", "example": "running" },
                    "message": { "type": "string", "example": "Job started. Poll GET /jobs/job_1738850000000_1 for status." }
                  }
                }
              }
            }
          },
          "400": { "description": "Campos obrigatórios faltando" },
          "401": { "description": "Token inválido" },
          "403": { "description": "Ferramenta não permitida" }
        }
      }
    },
    "/jobs/{jobId}": {
      "get": {
        "summary": "Consultar status de um job assíncrono",
        "description": "Retorna o status atual de um job. Se status='running', o job ainda está executando - faça polling a cada 3-5 segundos. Se status='completed' ou 'failed', o resultado final (stdout/stderr) está disponível.",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "ID do job retornado por POST /jobs/run",
            "example": "job_1738850000000_1"
          }
        ],
        "responses": {
          "200": {
            "description": "Status do job",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "id": { "type": "string" },
                    "tool": { "type": "string" },
                    "command": { "type": "string" },
                    "status": { "type": "string", "enum": ["running", "completed", "failed"] },
                    "stdout": { "type": "string", "description": "Saída do comando (vazio enquanto running)" },
                    "stderr": { "type": "string" },
                    "exitCode": { "type": "integer", "nullable": true },
                    "error": { "type": "string", "description": "Mensagem de erro (só em failed)" },
                    "elapsed": { "type": "integer", "description": "Tempo decorrido/total em ms" },
                    "createdAt": { "type": "integer", "description": "Timestamp de criação" },
                    "completedAt": { "type": "integer", "nullable": true, "description": "Timestamp de conclusão" }
                  }
                }
              }
            }
          },
          "404": { "description": "Job não encontrado (pode ter expirado após 1 hora)" }
        }
      }
    },
...
    "/files/read": {
      "post": {
        "summary": "Ler conteúdo de arquivo",
        "description": "Lê o conteúdo de um arquivo no PC remoto. Útil para ver código-fonte, configurações, logs, etc. Limite de 5MB por arquivo - para arquivos maiores, use /run com PowerShell (Get-Content).",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path"],
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Caminho absoluto do arquivo",
                    "example": "D:\\jarbas_vida\\package.json"
                  },
                  "encoding": {
                    "type": "string",
                    "description": "Encoding do arquivo (padrão: utf8)",
                    "default": "utf8",
                    "example": "utf8"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arquivo lido com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "content": { "type": "string", "description": "Conteúdo do arquivo" },
                    "size": { "type": "integer", "description": "Tamanho em bytes" },
                    "path": { "type": "string", "description": "Caminho absoluto resolvido" }
                  }
                }
              }
            }
          },
          "403": { "description": "Path traversal detectado ou fora dos roots permitidos" },
          "404": { "description": "Arquivo não encontrado" },
          "413": { "description": "Arquivo maior que 5MB" }
        }
      }
    },
    "/files/write": {
      "post": {
        "summary": "Criar ou escrever arquivo",
        "description": "Cria um novo arquivo ou sobrescreve um existente no PC remoto. Pode criar diretórios intermediários automaticamente com createDirs: true.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path", "content"],
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Caminho absoluto do arquivo",
                    "example": "D:\\projetos\\novo-arquivo.txt"
                  },
                  "content": {
                    "type": "string",
                    "description": "Conteúdo a escrever no arquivo",
                    "example": "Hello World!"
                  },
                  "createDirs": {
                    "type": "boolean",
                    "description": "Criar diretórios intermediários se não existirem",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Arquivo escrito com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "path": { "type": "string" },
                    "size": { "type": "integer", "description": "Tamanho do conteúdo escrito" },
                    "created": { "type": "boolean", "description": "true se arquivo foi criado, false se foi atualizado" }
                  }
                }
              }
            }
          },
          "400": { "description": "Campo content é obrigatório" },
          "403": { "description": "Path traversal detectado ou fora dos roots permitidos" }
        }
      }
    },
    "/files/list": {
      "post": {
        "summary": "Listar conteúdo de diretório",
        "description": "Lista arquivos e pastas de um diretório no PC remoto. Com recursive: true, lista até 3 níveis de profundidade.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["path"],
                "properties": {
                  "path": {
                    "type": "string",
                    "description": "Caminho absoluto do diretório",
                    "example": "D:\\jarbas_vida\\src"
                  },
                  "recursive": {
                    "type": "boolean",
                    "description": "Listar subdiretórios recursivamente (max 3 níveis)",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Diretório listado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "path": { "type": "string" },
                    "entries": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/FileEntry" }
                    },
                    "count": { "type": "integer", "description": "Número de entradas no nível raiz" }
                  }
                }
              }
            }
          },
          "403": { "description": "Path traversal detectado ou fora dos roots permitidos" },
          "404": { "description": "Diretório não encontrado" }
        }
      }
    },
...
  }
}